fix(delivery): complete Alpha.3 source reconciliation
Split KFD runtime acceptance coverage and accept successful source-bridge results on the clean release-cut successor.
