+++ .github/workflows/dev-pr-auto-merge.yml
@@ -790,12 +790,14 @@ jobs:
             > "$source_jobs"
           jq -e \
             --arg deliveryClass "$DELIVERY_CLASS" \
+            --arg priorRunConclusion "$prior_run_conclusion" \
             '[.jobs[] | select(.name == "Provisional Delivery Warrant admission / linux")]
              | select(length == 1)
              | .[0]
              | select(
                  .status == "completed" and
                  (.conclusion == "success" or
+                  ($priorRunConclusion == "success" and .conclusion == "skipped") or
                   ($deliveryClass == "source-only" and .conclusion == "skipped"))
                )' \
             "$source_jobs" >/dev/null
