+++ framework/release/publication-surfaces.json
@@ -198,7 +198,7 @@
           "classification": "non-publication",
           "owner": "kungfu-ci",
           "surfaceIds": [],
-          "sourceRoot": "sha256:1a7094d615dca215e9202f56da7a0cd3577cbb81f8f61b57f568684b37b1a5f7"
+          "sourceRoot": "sha256:f2c777415f61cd9b543dd04ad60acd9f0245378ac3aadab78f50cab1b9a22383"
         },
         {
           "path": ".github/workflows/dev-gate-latency-patrol.yml",
@@ -877,5 +877,5 @@
       }
     }
   ],
-  "registryRoot": "sha256:bce1c77418cbe26bdd8b44e3f146435d4ae9dfc76c3e2bbcfd4786a4eb9e038a"
+  "registryRoot": "sha256:ab8a3b91d075325503af192c89b8cc30b34ecf23ba0065dd356e7c27f44b9a02"
 }
